Menu, HUD, de todo un poco

+++ ProyectoII/Enemigo.h
@@ -14,7 +14,7 @@ public:
 	~Enemigo();
 	virtual void update();
 	inline void restaVida() { vida--; };
-	inline int getVida()const { return vida; };
+	inline float getVida()const { return vida; };
 
 private:
 	MundoVirtual* pMundo;
